requirements.txt
Flask~=2.3.2
line-bot-sdk
PyMySQL~=1.0.3
DBUtils~=3.0.3
loguru~=0.7.0